splice_experiment/requirements.txt
# Core dependencies (should already be in parent project)
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0

# For analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0

